+++ src/governance/index.js
@@ -1,0 +1,1 @@
+export { BallotBox, CONSTITUTIONAL_SOURCE } from "./ballots.js";